Added menus
Note: Currently it's not possible to pass data into the Win32 thread
that the window was created from. Because of thread safety issues
with Win32, we can't pass any data into the create_callback function.

Therefore, a potential user has to - right now - edit azul itself to
add the menus. Later on, menus should be replaced by code that azul
can control, so that we don't need to rely on Win32 to draw menus for
us.

This is a temporary solution and therefore has its own branch. Any
changes on master should also be ported to this branch.